refactor: DispatchingResolver::from_config() を追加し呼び出し元を一元化
言語ごとの Resolver 構築ロジックを DispatchingResolver::from_config() に集約。
新言語追加時の変更箇所を resolver/mod.rs の 1 ファイルに限定する。
副次的に pypi の TS aliases 未対応バグも修正。
